requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
django-hosts==6.0
django-subdomains==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
include==0.2.2
mysqlclient==2.2.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
django-cors-headers==3.7.0